feat: add slideshow carousel for Red Kuffiyeh AF1 images
Introduce image carousel with navigation and transitions.

#VERCEL_SKIP
